shinylive: Run 'shiny' Applications in the Browser

Exporting 'shiny' applications with 'shinylive' allows you to run them entirely in a web browser, without the need for a separate R server. The traditional way of deploying 'shiny' applications involves in a separate server and client: the server runs R and 'shiny', and clients connect via the web browser. When an application is deployed with 'shinylive', R and 'shiny' run in the web browser (via 'webR'): the browser is effectively both the client and server for the application. This allows for your 'shiny' application exported by 'shinylive' to be hosted by a static web server.

See also

Author

Maintainer: Barret Schloerke barret@posit.co (ORCID)

Authors:

Winston Chang winston@posit.co (ORCID)

George Stagg george.stagg@posit.co

Garrick Aden-Buie garrick@posit.co (ORCID)

Other contributors:

Posit Software, PBC (03wc8by49) [copyright holder, funder]